backend/pyproject.toml
[project]
name = "multi-agent-cae-v0-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[all]>=0.115.13",
    "google-adk>=1.5.0",
    "pydantic>=2.11.7",
]
